start making hourly fragment

+++ app/src/main/java/com/denbondd/justweather/models/currentweatherowm/Wind.java
@@ -6,12 +6,12 @@ import com.google.gson.annotations.SerializedName;
 public class Wind {
     @SerializedName("speed")
     @Expose
-    private Integer speed;
+    private Double speed;
     @SerializedName("deg")
     @Expose
     private Integer deg;
 
-    public Integer getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
